pyproject.toml
[tool.poetry]
name = "babydog"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1"
black = "^24.8.0"
isort = "^5.13.2"
django-allauth = "^64.0.0"
requests = "^2.32.3"
jwt = "^1.3.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
pyjwt = "^2.9.0"
django-environ = "^0.11.2"
simplejwt = "^2.0.1"
pillow = "^10.4.0"
environ = "^1.0"
django-cors-headers = "^4.4.0"
dj-rest-auth = "^6.0.0"
djangorestframework-simplejwt = "^5.3.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"